resource-agent/requirements.txt
# Resource Agent Dependencies
# Cross-platform support for Windows and Linux

# WebSocket client
websockets>=12.0

# HTTP client for REST calls
httpx>=0.27.0

# System information (cross-platform)
psutil>=5.9.0

# YAML config parsing
pyyaml>=6.0

# Async support
anyio>=4.0.0

# Data validation
pydantic>=2.0.0

# Logging
structlog>=24.0.0
